object pool

+++ Assets/Scripts/Runtime/Weapons/Data/WeaponConfiguration.cs
@@ -6,7 +6,7 @@ namespace TopDos.Weapons.Data
     public class WeaponConfiguration : ScriptableObject
     {
         [field: SerializeField] public string Name { get; private set; }
-        [field: SerializeField] public float Damage { get; private set; }
+        [field: SerializeField] public int Damage { get; private set; }
         [field: SerializeField] public float FireRate { get; private set; }
         [field: SerializeField] public float MaxAmmo { get; private set; }
     }
